+++ core/src/main/scala/com/krux/hyperion/DataPipelineDefWrapper.scala
@@ -12,6 +12,7 @@ case class DataPipelineDefWrapper private[hyperion] (
   override val hc: HyperionContext,
   override val pipelineName: String,
   schedule: Schedule,
+  override val pipelineLifeCycle: PipelineLifeCycle,
   workflowFunc: () => WorkflowExpression,  // for delayed workflow execution
   override val tags: Map[String, Option[String]],
   override val parameters: Iterable[Parameter[_]]
@@ -19,6 +20,7 @@ case class DataPipelineDefWrapper private[hyperion] (
 
   def withName(name: String) = copy(pipelineName = name)
   def withSchedule(schedule: Schedule) = copy(schedule = schedule)
+  def withPipelineLifeCycle(pipelineLifeCycle: PipelineLifeCycle) = copy(pipelineLifeCycle = pipelineLifeCycle)
   def withTags(tags: Map[String, Option[String]]) = copy(tags = this.tags ++ tags)
   def withParameters(parameters: Iterable[Parameter[_]]) = copy(parameters = parameters)
 
@@ -32,6 +34,7 @@ object DataPipelineDefWrapper {
     inner.hc,
     inner.pipelineName,
     inner.schedule,
+    inner.pipelineLifeCycle,
     () => inner.workflow,
     inner.tags,
     inner.parameters